+++ apps/native-component-list/src/screens/UI/SwitchScreen.tsx
@@ -9,11 +9,9 @@ export default function SwitchScreen() {
     <Page>
       <Section title="Switch">
         <Switch
-          checked={checked}
+          value={checked}
           style={{ width: 300, height: 50 }}
-          onCheckedChange={({ nativeEvent: { checked } }) => {
-            setChecked(checked);
-          }}
+          onValueChange={setChecked}
           color="#ff0000"
           label="Never gonna give you up"
           variant="switch"
@@ -21,11 +19,9 @@ export default function SwitchScreen() {
       </Section>
       <Section title="Checkbox Switch">
         <Switch
-          checked={checked}
+          value={checked}
           style={{ width: 300, height: 50 }}
-          onCheckedChange={({ nativeEvent: { checked } }) => {
-            setChecked(checked);
-          }}
+          onValueChange={setChecked}
           label="Never gonna let you down"
           color="#ff0000"
           variant="checkbox"
@@ -33,11 +29,9 @@ export default function SwitchScreen() {
       </Section>
       <Section title="Button Switch">
         <Switch
-          checked={checked}
+          value={checked}
           style={{ width: 300, height: 50 }}
-          onCheckedChange={({ nativeEvent: { checked } }) => {
-            setChecked(checked);
-          }}
+          onValueChange={setChecked}
           label="Never gonna run around and desert you"
           variant="button"
         />