+++ packages/expo-ui/ios/SwitchView.swift
@@ -4,11 +4,11 @@ import SwiftUI
 import ExpoModulesCore
 
 class SwitchProps: ExpoSwiftUI.ViewProps {
-  @Field var checked: Bool
+  @Field var value: Bool
   @Field var variant: String?
   @Field var label: String?
   @Field var color: Color?
-  var onCheckedChanged = EventDispatcher()
+  var onValueChange = EventDispatcher()
 }
 
 struct SwitchView: ExpoSwiftUI.View {
@@ -20,16 +20,16 @@ struct SwitchView: ExpoSwiftUI.View {
     ExpoSwiftUI.AutoSizingStack(shadowNodeProxy: shadowNodeProxy, axis: .both) {
       Toggle(isOn: $checked, label: { props.label != nil ? Text(props.label ?? "") : nil })
       .onChange(of: checked, perform: { newValue in
-        if props.checked == newValue {
+        if props.value == newValue {
           return
         }
-        props.onCheckedChange([
-          "checked": newValue
+        props.onValueChange([
+          "value": newValue
         ])
       })
       .tint(props.color)
       .onReceive(props.objectWillChange, perform: {
-        checked = props.checked
+        checked = props.value
       })
       .onAppear {
         checked = props.checked